Split profile part of tui_configure into separate file
